third_party/blink/web_tests/external/wpt/css/css-transitions/properties-value-implicit-001-expected.txt

This is a testharness.js-based test.
Found 1 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] background-position length-em(em) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "background-position:2s" but got "background-position-x:2s"
Harness: the test ran to completion.
